nft: Extract `config` to its own package
Separating the config implementation from the top level `nft` package,
which is used as the main API, enables the backend implementation/s
(exec for now) to be extracted as well to its own package.
