+++ nft/config/table_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package nft_test
+package config_test
 
 import (
 	"fmt"
